+++ ext/monitor/src/test/java/app/monitor/job/KubeMonitorJobTest.java
@@ -64,14 +64,14 @@ class KubeMonitorJobTest {
         pod.metadata.creationTimestamp = startTime;
 
         assertThat(job.check(pod, startTime.plusMinutes(1))).isNull();
-        assertThat(job.check(pod, startTime.plusMinutes(5))).isEqualTo("pod is still not ready, elapsed=PT5M");
+        assertThat(job.check(pod, startTime.plusMinutes(5))).isEqualTo("pod is not in ready state, uptime=PT5M");
 
         pod = pod("Running");
         pod.status.containerStatuses.get(0).ready = Boolean.FALSE;
         pod.status.startTime = startTime;
 
         assertThat(job.check(pod, startTime.plusMinutes(1))).isNull();
-        assertThat(job.check(pod, startTime.plusMinutes(5))).isEqualTo("pod is still not ready, elapsed=PT5M");
+        assertThat(job.check(pod, startTime.plusMinutes(5))).isEqualTo("pod is not in ready state, uptime=PT5M");
     }
 
     @Test